% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.to.R
\name{array.to}
\alias{array.to}
\title{Array to ...}
\usage{
array.to(array, to)
}
\arguments{
\item{array}{A \code{"gpagen"} object or an array of landmarks}

\item{to}{Either \code{"list"} or \code{"matrix"}.}
}
\description{
Converts a \code{"gpagen"} array to a \code{"list"} or a \code{"matrix"}
}
\examples{
## Loading the plethodon dataset
require(geomorph)
data(plethodon)

## Performing a procrustes superimposition
procrustes <- geomorph::gpagen(plethodon$land, print.progress = FALSE)

## Transforming it into a list
array.to(procrustes, "list")

## Transforming it into a matrix
array.to(procrustes, "matrix")


}
\author{
Thomas Guillerme
}
