% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landvR-package.R
\docType{package}
\name{landvR-package}
\alias{landvR}
\alias{landvR-package}
\title{Utility package for studying landmark variation}
\description{
This package contains utility functions to study landmark variations in a Procrustes space
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TGuillerme/landvR}
}

}
\author{
Thomas Guillerme <guillert@tcd.ie>
}
\keyword{disparity,}
\keyword{geometric}
\keyword{landmark,}
\keyword{morphometric,}
\keyword{permutation}
\keyword{random}
\keyword{test}
\keyword{test,}
