% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.dist.R
\name{linear.dist}
\alias{linear.dist}
\title{Linear distance}
\usage{
linear.dist(landmarks, measurements, dist = "euclidean")
}
\arguments{
\item{landmarks}{A \code{"matrix"}, \code{"list"} or \code{"array"}, where the landmarks are rows and the columns are dimensions.}

\item{measurements}{A pair of landmark IDs or a \code{"list"} of pairs of landmarks IDs.}

\item{dist}{The type of distance as passed to \code{\link[stats]{dist}} (default is \code{"euclidean"}).}
}
\description{
Measure the linear distance between pairs of landmarks
}
\examples{
## Loading the plethodon dataset
require(geomorph)
data(plethodon)
data(scallops)

## One measurement on multiple specimen
linear.dist(plethodon$land, c(1,2))

## List of multiple measurements on multiple specimen
linear.dist(scallops$coorddata, list("a" = c(1,2), "b" = c(2,15)))


}
\author{
Thomas Guillerme
}
