% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.var.plot.R
\name{procrustes.var.plot}
\alias{procrustes.var.plot}
\title{Plot Procrustes variation}
\usage{
procrustes.var.plot(M1, M2, col = list("grey", "black"), col.val,
  magnitude = 1, labels = FALSE, axes = FALSE, pt.size = 1,
  gridPar, col.range, ...)
}
\arguments{
\item{M1}{Matrix of landmark coordinates for the first (reference) specimen}

\item{M2}{Matrix of landmark coordinates for the second (target) specimen}

\item{col}{Either a single color value (\code{"character"}), vector of values or function for colouring both points and vectors; or a list of two of any of these three elements. The first argument is passed to the points and the second to the vectors (\code{default = list("grey", "black")} for grey points and black vectors - see details).}

\item{col.val}{Optional, if \code{col} is a function for assigning colours (or a list of two functions), which values to pass to the function to assign the colors).}

\item{magnitude}{The magnitude of difference between the two shapes (\code{default = 1}).}

\item{labels}{Logical, whether to plot the landmark labels (\code{TRUE}) or not (\code{FALSE} - default).}

\item{axes}{Logical, whether to plot the plot axes (\code{TRUE}) or not (\code{FALSE} - default).}

\item{pt.size}{The size of the points (\code{default = 1}).}

\item{gridPar}{Optional, a named list of grid parameters generated by \code{\link[geomorph]{gridPar}}.}

\item{col.range}{Optional, two \code{numeric} values for designating the minimum and maximum range to sample the colors from (see details). If left empty (default), the range of \code{col.val} is used.}

\item{...}{Additional parameters to \code{\link[graphics]{plot}} or \code{\link[rgl]{plot3d}}.}
}
\description{
Modification of the \code{\link[geomorph]{plotRefToTarget}} function to allow to display landmark variation on Procrustes
}
\details{
\code{col} arguments are passed to \code{\link[graphics]{plot}} or \code{\link[rgl]{plot3d}} as follows:
\itemize{
   \item if col is a single value or a single vector, it passes it to both points and colors arguments;
   \item if col is a single function, it applies it to both points and colors arguments based on the argument \code{col.val};
   \item if col is a list of two values, vectors or functions, it will apply the first element to the points and the second to the vectors the same way as described above;
}

If \code{col.val} is provided the function samples colors according the the \code{col} argument (see above) using the range of the \code{col.val} vector.
You can change this range to some absolute values by using the \code{col.range} option: the colors are then sampled along the \code{col.range} range for each \code{col.val}.
}
\examples{
## Loading the geomorph dataset
require(geomorph)
data(plethodon)

## Performing the Procrustes superimposition
proc_super <- gpagen(plethodon$land, print.progress = FALSE)

## Getting the range of variation
variation <- variation.range(proc_super, return.ID = TRUE)

## Selecting the coordinates and the variation vector
M1 <- proc_super$coords[, , variation$min.max[1]]
M2 <- proc_super$coords[, , variation$min.max[2]]
var_val <- variation$range[, 1]
 
## Plot the variation
procrustes.var.plot(M1, M2)

## A colourful plot
procrustes.var.plot(M1, M2, col = list(grDevices::rainbow, "pink"), col.val = var_val,
                    pt.size = 2.5)

## A series of heat plots with different options
par(mfrow = c(3,1), mar = c(0,0,2,0))
procrustes.var.plot(M1, M2, main = "Uncolored variation")
procrustes.var.plot(M1, M2, col = grDevices::heat.colors, col.val = var_val, 
                    main = "Relative colours")
procrustes.var.plot(M1, M2, col = grDevices::heat.colors, col.val = var_val,
                    col.range = c(0, 0.2), main = "Absolute colours (range = c(0, 0.2))")

\dontrun{
## Loading the scallops 3D data from geomorph
require(geomorph)
data(scallops)

## Procrustes superimposition
procrustes <- gpagen(scallops$coorddata)

## Getting the range of variation
variation <- variation.range(procrustes, return.ID = TRUE)

## Selecting the coordinates and the variation vector
M1 <- procrustes$coords[, , variation$min.max[1]]
M2 <- procrustes$coords[, , variation$min.max[2]]
var_val <- variation$range[, 1]
 
## Plot the variation in 3D
procrustes.var.plot(M1, M2, col.val = var_val, col = list(grDevices::heat.colors, "grey"))
}

}
\seealso{
\code{\link{variation.range}}
}
\author{
Thomas Guillerme - original function by Dean Adams, Emma Sherratt & Michael Collyer.
}
