% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.part.names.R
\name{prop.part.names}
\alias{prop.part.names}
\title{Get the partitions tip labels from a tree}
\usage{
prop.part.names(phy, singletons = FALSE, inc.nodes = FALSE)
}
\arguments{
\item{phy}{an object of class \code{"phylo"}}

\item{singletons}{logical, whether to include tips as singletons (\code{TRUE}) or not (\code{FALSE} - default)}

\item{inc.nodes}{logical, whether to include nodes in the clades (\code{TRUE}) or not (\code{FALSE} - default). If \code{TRUE}, \code{phy} must have node labels.}
}
\description{
Get the names of the tip labels in each bipartition (clade) of a tree.
}
\examples{
## The partition's tip labels
prop.part.names(ape::rtree(5))

## The partition's tip labels (including singletons)
prop.part.names(ape::rtree(3), singletons = TRUE)

## The partition's tip labels of a list of trees
lapply(ape::rmtree(2, 3), prop.part.names)

}
\seealso{
\code{\link[ape]{prop.part}}
}
\author{
Thomas Guillerme
}
