% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.procrustes.R
\name{select.procrustes}
\alias{select.procrustes}
\title{Select Procrustes values}
\usage{
select.procrustes(procrustes, selector = mean, factors, specimen = FALSE)
}
\arguments{
\item{procrustes}{Procrustes data of class \code{"gpagen"} or an \code{array} of coordinates.}

\item{selector}{A \code{function} of which values to select (default = \code{\link{mean}}).}

\item{factors}{A \code{list} of elements names or IDs to split the data.}

\item{specimen}{Whether to return the values of the estimated selector (\code{FALSE} - default - a non-existing specimen) or the specimen the closest to the the estimated selector (\code{TRUE}).}
}
\value{
A procrustes value

A specimen number
}
\description{
Selects specific values of specimens from a Procrustes superimposition
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
