% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.diff.R
\name{slope.diff}
\alias{slope.diff}
\title{Slope difference}
\usage{
slope.diff(slope1, slope2, angle = "degree", significance)
}
\arguments{
\item{slope1}{A slope.}

\item{slope2}{Another slope.}

\item{angle}{Either \code{"degree"} (default) or \code{"radian"}.}

\item{significance}{Optional, an angle value below which angles are not significant (in degree or radian)}
}
\value{
Either the value of the angle (if \code{significance} is left empty) or a \code{logical} whether the angle is below the \code{significance} value.
}
\description{
Measuring the difference between slopes in angles
}
\examples{
## A random dataset
data <- as.data.frame(replicate(3, rnorm(10)))
colnames(data) <- letters[1:3]

## Two slopes
slope_b <- lm(a ~ b, data)$coefficients[["b"]]
slope_c <- lm(a ~ c, data)$coefficients[["c"]]

## The slope difference
slope.diff(slope_b, slope_c)
## The slope difference in radian
slope.diff(slope_b, slope_c, angle = "radian")
## Whether the slope difference is lower than 4.5 degrees
slope.diff(slope_b, slope_c, significance = 4.5)

}
\author{
Thomas Guillerme and Ariel Marcy
}
