% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.part.names.R
\name{prop.part.names}
\alias{prop.part.names}
\title{Get the partitions tip labels from a tree}
\usage{
prop.part.names(phy, singletons = FALSE)
}
\arguments{
\item{phy}{an object of class \code{"phylo"}}

\item{singletons}{logical, whether to include tips as singletons (\code{TRUE}) or not (\code{FALSE} - default)}
}
\description{
Get the names of the tip labels in each bipartition (clade) of a tree.
}
\examples{
## The partition's tip labels
prop.part.names(rtree(5))

## The partition's tip labels (including singletons)
prop.part.names(rtree(3), singletons = TRUE)

## The partition's tip labels of a list of trees
lapply(rmtree(2, 3), prop.part.names)

}
\seealso{
\code{\link[ape]}{prop.part}
}
\author{
Thomas Guillerme
}
