% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.procrustes.R
\name{select.procrustes}
\alias{select.procrustes}
\title{Select Procrustes values}
\usage{
select.procrustes(procrustes, selector = mean, factors)
}
\arguments{
\item{procrustes}{Procrustes data of class \code{"gpagen"} or an \code{"array"} or \code{"list"} of coordinates.}

\item{selector}{A \code{function} of which values to select (default = \code{\link{mean}}).}

\item{factors}{A \code{list} of elements names or IDs to split the data.}
}
\value{
The coordinates of a hypothetical specimen (e.g. the mean specimen).
}
\description{
Selects specific values of specimens from a Procrustes superimposition
}
\examples{
## Loading the plethodon dataset
require(geomorph)
data(plethodon)

## Performing a procrustes superimposition
procrustes <- geomorph::gpagen(plethodon$land, print.progress = FALSE)

## Selecting the mean Procrustes
mean_procrustes <- select.procrustes(procrustes, selector = mean)

## Selecting the minimum Procrustes shape for each species
min_procrustes <- select.procrustes(procrustes, selector = min,
                                    factors = list(which(plethodon$species == "Jord"),
                                                   which(plethodon$species == "Teyah")))

}
\seealso{

}
\author{
Thomas Guillerme
}
